project/pyproject.toml
[project]
name = "order-dashboard"
version = "1.0.0"
description = "Order Management Dashboard with Flask and MySQL"
authors = [{ name = "Modern SWE Team" }]
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "mysql-connector-python>=8.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-flask>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
